refactor: adjust avg_rating field type for store

+++ src/main/java/nbcamp/TwoFastDelivery/domain/store/application/dto/StoreDetailResponse.java
@@ -21,6 +21,6 @@ public class StoreDetailResponse {
     private final String closeTime;
     private final String description;
     private final String status;
-    private final Byte avgRating;
+    private final Float avgRating;
     private final Integer reviewCount;
 }
